chikitsa_backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
